requirements.txt
numpy==1.25.2
einops==0.6.1
h5py==3.9.0
matplotlib-label-lines==0.6.0
seaborn==0.12.2
wandb==0.15.7
PyYAML==6.0.1
argparse==1.4.0
tqdm==4.65.0
matplotlib==3.7.2
pandas==2.0.3
scipy==1.11.1
hydra-core==1.3.2
torch-intermediate-layer-getter==0.1.post1
scikit-learn==1.3.0